Fix length mask crc32c

+++ Sources/SwiftTFRecords/Helper.swift
@@ -11,8 +11,7 @@ func maskCrc(_ crc: UInt32) -> UInt32 {
     UInt32(Int(Int((crc >> 15) | (crc << 17)) + 0xa282ead8) % Int(pow(2.0, 32.0)))
 }
 
-func intToArray(_ n: UInt32) -> [UInt8] {
+func intToArray(_ n: UInt64) -> [UInt8] {
     let data = withUnsafeBytes(of: n) { Data($0) }
     return [UInt8](data)
 }
-
